修复解析geometry时如果出错并设置忽略错误时应该返回[]byte{}

+++ mysql-binlog/analyzer.go
@@ -167,7 +167,7 @@ func (a *analyzer) parseValue(t int, rawType string, v interface{}) (interface{}
 				p, err := a.parseWKB(raw)
 				if err != nil && a.IgnoreWKBDataParseError {
 					a.app.Warn("parse geometry data error", zap.Error(err))
-					return []float64{}, nil
+					return []byte{}, nil
 				}
 				return p, err
 			}
